requirements.txt
async-timeout==4.0.2
black==23.1.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
filelock==3.10.0
flake8==6.0.0
identify==2.5.21
idna==3.4
isort==5.12.0
mccabe==0.7.0
mongoengine==0.27.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pre-commit==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.3.3
python-dateutil==2.8.2
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
spotipy==2.22.1
tomli==2.0.1
urllib3==1.26.14
virtualenv==20.21.0
